Trimester-4/MICROSERVICES/CIA3/services/gateway_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.22.0
strawberry-graphql[fastapi]>=0.209.3
httpx>=0.27.0
pydantic-settings>=2.2
python-dotenv>=1.0
